Bug 2020608 - Part 15: Require unsigned integral types for CeilingLog2 and FindMostSignificantBit. r=sergesanspaille,nbp
Instead of implicitly casting the input to unsigned, require this happens explicitly
in the caller. This aligns with `<bit>` std-functions and `Rotate{Left,Right}` which
also require unsigned inputs.

Differential Revision: https://phabricator.services.mozilla.com/D286324

UltraBlame original commit: 8fc5677faae40170e1f846f6ea0578ee0801481d

+++ js/src/jit/MacroAssembler.cpp
@@ -2064,7 +2064,7 @@ void MacroAssembler::loadInt32ToStringWithBase(
   {
     
     if (std::has_single_bit(uint32_t(base))) {
-      uint32_t shift = mozilla::FloorLog2(base);
+      uint32_t shift = mozilla::FloorLog2(uint32_t(base));
 
       rshift32(Imm32(shift), input, scratch1);
       and32(Imm32((uint32_t(1) << shift) - 1), input, scratch2);
