+++ js/src/jit/loong64/CodeGenerator-loong64.cpp
@@ -797,7 +797,7 @@ void CodeGenerator::visitMulI(LMulI* ins) {
     }
 
     if (constant > 0) {
-      uint32_t shift = mozilla::FloorLog2(constant);
+      uint32_t shift = mozilla::FloorLog2(uint32_t(constant));
 
       if (!mul->canOverflow()) {
         
@@ -899,7 +899,7 @@ void CodeGeneratorLOONG64::emitMulI64(Register lhs, int64_t rhs,
 
   if (rhs > 0) {
     if (std::has_single_bit(static_cast<uint64_t>(rhs + 1))) {
-      int32_t shift = mozilla::FloorLog2(rhs + 1);
+      int32_t shift = mozilla::FloorLog2(uint64_t(rhs + 1));
 
       UseScratchRegisterScope temps(masm);
       Register savedLhs = lhs;
@@ -913,7 +913,7 @@ void CodeGeneratorLOONG64::emitMulI64(Register lhs, int64_t rhs,
     }
 
     if (std::has_single_bit(static_cast<uint64_t>(rhs - 1))) {
-      int32_t shift = mozilla::FloorLog2(rhs - 1);
+      int32_t shift = mozilla::FloorLog2(uint64_t(rhs - 1));
       if (shift < 5) {
         masm.as_alsl_d(dest, lhs, lhs, shift - 1);
       } else {
@@ -930,7 +930,7 @@ void CodeGeneratorLOONG64::emitMulI64(Register lhs, int64_t rhs,
     }
 
     
-    int32_t shift = mozilla::FloorLog2(rhs);
+    int32_t shift = mozilla::FloorLog2(uint64_t(rhs));
     if (int64_t(1) << shift == rhs) {
       masm.as_slli_d(dest, lhs, shift);
       return;
