+++ js/src/wasm/WasmBCStkMgmt-inl.h
@@ -1154,7 +1154,7 @@ bool BaseCompiler::popConstPositivePowerOfTwo(int32_t* c, uint_fast8_t* power,
   if (*c <= cutoff || !std::has_single_bit(static_cast<uint32_t>(*c))) {
     return false;
   }
-  *power = mozilla::FloorLog2(*c);
+  *power = mozilla::FloorLog2(uint32_t(*c));
   stk_.popBack();
   return true;
 }
@@ -1169,7 +1169,7 @@ bool BaseCompiler::popConstPositivePowerOfTwo(int64_t* c, uint_fast8_t* power,
   if (*c <= cutoff || !std::has_single_bit(static_cast<uint64_t>(*c))) {
     return false;
   }
-  *power = mozilla::FloorLog2(*c);
+  *power = mozilla::FloorLog2(uint64_t(*c));
   stk_.popBack();
   return true;
 }
